pyproject.toml
[tool.poetry]
name = "opennempy"
version = "3.6.0-alpha.1"
description = "OpenNEM Australian Energy Data Python Client"
authors = ["Dylan McConnell <dylan.mcconnell@unimelb.edu.au>", "Nik Cubrilovic <nik@infotorch.org>"]
license = "MIT"

readme = "README.md"

homepage = "https://opennem.org.au"
repository = "https://github.com/opennem/opennempy"

keywords = ["opennem", "analysis", "australia"]

classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Framework :: Scrapy",
  "License :: OSI Approved :: MIT License",
]

include = [
]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
pydantic = "^1.5.1"
click = "^7.1.2"
requests_cache = "^0.5.2"
python-dateutil = "^2.8.1"
python-dotenv = "^0.15.0"
pandas = "^1.2.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.7.8"
isort = "^5.6.4"
pylint = "^2.2"
rope = "^0.18.0"
twine = "^3.1.1"
ipython = "^7.15.0"
pytest = "^6.1.2"
pytest-watch = "^4.2.0"
mypy = "^0.790"
pycodestyle = "^2.6.0"
py-spy = "^0.3.3"
pytest-benchmark = {extras = ["histogram"], version = "^3.2.3"}
pdbpp = "^0.10.2"

[tool.poetry.scripts]
opennem = 'opennem.cli:main'

[tool.black]
line-length = 99
target_version = ['py38']
include = '\.pyi?$'
exclude = '''
/(\.git/
  |\.eggs
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 99

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
